add moving average option to TradeProfitLoss Widget
in addition to the existing FIFO method
